+++ packages/modernjs/src/cli/utils.ts
@@ -7,7 +7,7 @@ import { LOCALHOST, PLUGIN_IDENTIFIER } from '../constant';
 import logger from './logger';
 import { autoDeleteSplitChunkCacheGroups } from '@module-federation/rsbuild-plugin/utils';
 
-import type { BundlerConfig, BundlerChainConfig } from '../interfaces/bundler';
+import type { BundlerChainConfig } from '../interfaces/bundler';
 import type {
   webpack,
   UserConfig,
@@ -208,21 +208,20 @@ export const patchMFConfig = (
   return mfConfig;
 };
 
-export function patchIgnoreWarning<T extends Bundler>(
-  bundlerConfig: BundlerConfig<T>,
-) {
-  bundlerConfig.ignoreWarnings = bundlerConfig.ignoreWarnings || [];
+function patchIgnoreWarning<T extends Bundler>(chain: BundlerChainConfig) {
+  const ignoreWarnings = chain.get('ignoreWarnings') || [];
   const ignoredMsgs = [
     'external script',
     'process.env.WS_NO_BUFFER_UTIL',
     `Can't resolve 'utf-8-validate`,
   ];
-  bundlerConfig.ignoreWarnings.push((warning) => {
+  ignoreWarnings.push((warning) => {
     if (ignoredMsgs.some((msg) => warning.message.includes(msg))) {
       return true;
     }
     return false;
   });
+  chain.ignoreWarnings(ignoreWarnings);
 }
 
 export function addMyTypes2Ignored(
@@ -277,87 +276,78 @@ export function addMyTypes2Ignored(
     ignored: ignored.concat(DEFAULT_IGNORED_GLOB),
   });
 }
-export function patchBundlerConfig<T extends Bundler>(options: {
-  bundlerConfig: BundlerConfig<T>;
+export function patchBundlerConfig(options: {
+  chain: BundlerChainConfig;
   isServer: boolean;
   modernjsConfig: UserConfig<AppTools>;
-  bundlerType: Bundler;
   mfConfig: moduleFederationPlugin.ModuleFederationPluginOptions;
+  enableSSR: boolean;
 }) {
-  const { bundlerConfig, modernjsConfig, isServer, mfConfig, bundlerType } =
-    options;
-  const enableSSR = Boolean(modernjsConfig.server?.ssr);
+  const { chain, modernjsConfig, isServer, mfConfig, enableSSR } = options;
 
-  delete bundlerConfig.optimization?.runtimeChunk;
+  chain.optimization.delete('runtimeChunk');
 
-  patchIgnoreWarning(bundlerConfig);
+  patchIgnoreWarning(chain);
 
-  if (bundlerConfig.output) {
-    if (!bundlerConfig.output?.chunkLoadingGlobal) {
-      bundlerConfig.output.chunkLoadingGlobal = `chunk_${mfConfig.name}`;
-    }
-    if (!bundlerConfig.output?.uniqueName) {
-      bundlerConfig.output.uniqueName = mfConfig.name;
-    }
+  if (!chain.output.get('chunkLoadingGlobal')) {
+    chain.output.chunkLoadingGlobal(`chunk_${mfConfig.name}`);
+  }
+  if (!chain.output.get('uniqueName')) {
+    chain.output.uniqueName(mfConfig.name!);
   }
 
+  const splitChunkConfig = chain.optimization.splitChunks.entries();
   if (!isServer) {
-    // @ts-ignore
-    autoDeleteSplitChunkCacheGroups(mfConfig, bundlerConfig);
+    // @ts-ignore type not the same
+    autoDeleteSplitChunkCacheGroups(mfConfig, splitChunkConfig);
   }
 
   if (
     !isServer &&
     enableSSR &&
-    typeof bundlerConfig.optimization?.splitChunks === 'object' &&
-    bundlerConfig.optimization.splitChunks.cacheGroups
+    splitChunkConfig &&
+    typeof splitChunkConfig === 'object' &&
+    splitChunkConfig.cacheGroups
   ) {
-    bundlerConfig.optimization.splitChunks.chunks = 'async';
+    splitChunkConfig.chunks = 'async';
     logger.warn(
       `splitChunks.chunks = async is not allowed with stream SSR mode, it will auto changed to "async"`,
     );
   }
 
-  if (isDev && bundlerConfig.output?.publicPath === 'auto') {
+  if (isDev && chain.output.get('publicPath') === 'auto') {
     // TODO: only in dev temp
     const port =
       modernjsConfig.dev?.port || modernjsConfig.server?.port || 8080;
     const publicPath = `http://localhost:${port}/`;
-    bundlerConfig.output.publicPath = publicPath;
+    chain.output.publicPath(publicPath);
   }
 
   if (isServer && enableSSR) {
-    const { output } = bundlerConfig;
-    const uniqueName = mfConfig.name || output?.uniqueName;
-    const chunkFileName = output?.chunkFilename;
+    const uniqueName = mfConfig.name || chain.output.get('uniqueName');
+    const chunkFileName = chain.output.get('chunkFilename');
     if (
-      output &&
       typeof chunkFileName === 'string' &&
       uniqueName &&
       !chunkFileName.includes(uniqueName)
     ) {
       const suffix = `${encodeName(uniqueName)}-[chunkhash].js`;
-      output.chunkFilename = chunkFileName.replace('.js', suffix);
+      chain.output.chunkFilename(chunkFileName.replace('.js', suffix));
     }
   }
   // modernjs project has the same entry for server/client, add polyfill:false to skip compile error in browser target
   if (isDev && enableSSR && !isServer) {
-    bundlerConfig.resolve!.fallback = {
-      ...bundlerConfig.resolve!.fallback,
-      crypto: false,
-      stream: false,
-      vm: false,
-    };
+    chain.resolve.fallback
+      .set('crypto', false)
+      .set('stream', false)
+      .set('vm', false);
   }
 
   if (
     modernjsConfig.deploy?.microFrontend &&
     Object.keys(mfConfig.exposes || {}).length
   ) {
-    if (!bundlerConfig.optimization) {
-      bundlerConfig.optimization = {};
-    }
-    bundlerConfig.optimization.usedExports = false;
+    chain.optimization.usedExports(false);
   }
 }
 
@@ -390,34 +380,12 @@ export const getIPV4 = (): string => {
   return ipv4Interface.address;
 };
 
-// lib-polyfill.js: include core-js，@babel/runtime，@swc/helpers，tslib.
-// lib-react.js: include react，react-dom.
-// lib-router.js: include react-router，react-router-dom，history，@remix-run/router.
-// lib-lodash.js: include lodash，lodash-es.
-// lib-antd.js: include antd.
-// lib-arco.js: include @arco-design/web-react.
-// lib-semi.js: include @douyinfe/semi-ui.
-// lib-axios.js: include axios.
-
-const SPLIT_CHUNK_MAP = {
-  REACT: 'react',
-  ROUTER: 'router',
-  LODASH: 'lib-lodash',
-  ANTD: 'lib-antd',
-  ARCO: 'lib-arco',
-  SEMI: 'lib-semi',
-  AXIOS: 'lib-axios',
-};
-const SHARED_SPLIT_CHUNK_MAP = {
-  react: SPLIT_CHUNK_MAP.REACT,
-  'react-dom': SPLIT_CHUNK_MAP.REACT,
-  'react-router': SPLIT_CHUNK_MAP.ROUTER,
-  'react-router-dom': SPLIT_CHUNK_MAP.ROUTER,
-  '@remix-run/router': SPLIT_CHUNK_MAP.ROUTER,
-  lodash: SPLIT_CHUNK_MAP.LODASH,
-  'lodash-es': SPLIT_CHUNK_MAP.LODASH,
-  antd: SPLIT_CHUNK_MAP.ANTD,
-  '@arco-design/web-react': SPLIT_CHUNK_MAP.ARCO,
-  '@douyinfe/semi-ui': SPLIT_CHUNK_MAP.SEMI,
-  axios: SPLIT_CHUNK_MAP.AXIOS,
+export const isWebTarget = (target: string[] | string) => {
+  const WEB_TARGET = 'web';
+  if (Array.isArray(target)) {
+    return target.includes(WEB_TARGET);
+  } else if (typeof target === 'string') {
+    return target === WEB_TARGET;
+  }
+  return false;
 };