+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @module-federation/nextjs-mf
 
+## 8.8.18
+
+### Patch Changes
+
+- Updated dependencies [0f71cbc]
+- Updated dependencies [22fcccd]
+  - @module-federation/sdk@0.10.0
+  - @module-federation/enhanced@0.10.0
+  - @module-federation/node@2.6.28
+  - @module-federation/runtime@0.10.0
+  - @module-federation/webpack-bundler-runtime@0.10.0
+
 ## 8.8.17
 
 ### Patch Changes