+++ packages/node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @module-federation/node
 
+## 2.6.28
+
+### Patch Changes
+
+- Updated dependencies [0f71cbc]
+- Updated dependencies [22fcccd]
+  - @module-federation/sdk@0.10.0
+  - @module-federation/enhanced@0.10.0
+  - @module-federation/runtime@0.10.0
+  - @module-federation/utilities@3.1.46
+
 ## 2.6.27
 
 ### Patch Changes