+++ packages/runtime-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/runtime
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies [0f71cbc]
+- Updated dependencies [22fcccd]
+  - @module-federation/sdk@0.10.0
+  - @module-federation/error-codes@0.10.0
+
 ## 0.9.1
 
 ### Patch Changes